move `{commandline,config,utils}` to `common.{commandline,config,utils}`
