+++ commands/explain.py
@@ -1,7 +1,9 @@
+import random
+
+from common import utils
 from lrrbot import bot, log
 import storage
-import random
-import utils
+
 
 @bot.command("explain (.*?)")
 @utils.throttle(5, params=[4])
