+++ commands/highlights.py
@@ -1,9 +1,12 @@
+import time
+
+import irc.client
+
+from common import utils
 from lrrbot import bot
 import twitch
-import utils
-import time
 import storage
-import irc.client
+
 
 @bot.command("highlight (.*?)")
 @utils.sub_only
