+++ common/config.py
@@ -1,7 +1,10 @@
 import configparser
 import logging
+
 import pytz
-from commandline import argv
+
+from common.commandline import argv
+
 
 CONFIG_SECTION = 'lrrbot'
