+++ twitch.py
@@ -1,7 +1,9 @@
 import json
-import utils
+
+from common import utils
+from common.config import config
 import storage
-from config import config
+
 
 def get_info(username=None, use_fallback=True):
 	"""
