+++ www/archive.py
@@ -1,14 +1,18 @@
-import flask
-import flask.json
-from www import server
-from www import login
-import urllib.request, urllib.parse
+import urllib.request
+import urllib.parse
 import time
 import os
-import dateutil.parser
-import utils
 import contextlib
 import tempfile
+
+import flask
+import flask.json
+import dateutil.parser
+
+from common import utils
+from www import server
+from www import login
+
 
 CACHE_TIMEOUT = 15*60
